Fix mypy and pre-commit hooks
