+++ .pre-commit-config.yaml
@@ -9,16 +9,6 @@ repos:
     rev: 1.16.0
     hooks:
       - id: yamlfix
-  - repo: local
-    hooks:
-      - id: check-nbqa-version-mismatch
-        name: Check for version mismatch between black, ruff, and nbQA
-        entry: python scripts/check_nbqa_version_mismatch.py
-        language: python
-        always_run: true
-        require_serial: true
-        additional_dependencies:
-          - pyyaml
   - repo: https://github.com/pre-commit/pre-commit-hooks
     rev: v4.6.0
     hooks:
@@ -55,30 +45,27 @@ repos:
     rev: v1.35.1
     hooks:
       - id: yamllint
-  - repo: https://github.com/psf/black
-    rev: 24.4.2
-    hooks:
-      - id: black
-        language_version: python3.12
   - repo: https://github.com/asottile/blacken-docs
     rev: 1.18.0
     hooks:
       - id: blacken-docs
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: v0.5.0
+    rev: v0.5.1
     hooks:
+      # Run the linter.
       - id: ruff
-      #   args:
-      #     - --verbose
-  - repo: https://github.com/nbQA-dev/nbQA
-    rev: 1.8.5
-    hooks:
-      - id: nbqa-black
-        additional_dependencies:
-          - black==24.4.2
-      - id: nbqa-ruff
-        additional_dependencies:
-          - ruff==v0.5.0
+        types_or:
+          - python
+          - pyi
+          - jupyter
+        args:
+          - --fix
+      # Run the formatter.
+      - id: ruff-format
+        types_or:
+          - python
+          - pyi
+          - jupyter
   - repo: https://github.com/executablebooks/mdformat
     rev: 0.7.17
     hooks:
