+++ pyproject.toml
@@ -164,6 +164,7 @@ test-gpu = {features = ["test", "cuda"], solve-group = "cuda"}
 [tool.ruff]
 target-version = "py312"
 fix = true
+exclude = ["src/lcm/sandbox"]
 
 [tool.ruff.lint]
 select = ["ALL"]
@@ -215,6 +216,9 @@ extend-ignore = [
 
     # long messages outside the exception class
     "TRY003",
+
+    # Missing docstring in magic method
+    "D105",
 ]
 
 [tool.ruff.lint.per-file-ignores]
@@ -223,6 +227,7 @@ extend-ignore = [
 "examples/*" = ["INP001"]
 "explanations/*" = ["INP001", "B018", "T201", "E402", "PD008"]
 "scripts/*" = ["INP001", "D101", "RET503"]
+"**/*.ipynb" = ["FBT003", "E402"]
 
 [tool.ruff.lint.pydocstyle]
 convention = "google"
@@ -238,7 +243,6 @@ black = "pyproject.toml"
 black = 1
 
 [tool.nbqa.exclude]
-ruff = "src/lcm/sandbox"
 black = "src/lcm/sandbox"
 
 # ======================================================================================
