+++ src/lcm/__init__.py
@@ -1,5 +1,4 @@
 from lcm import mark
-from lcm.user_grids import DiscreteGrid, LinspaceGrid, LogspaceGrid
-from lcm.user_model import Model
+from lcm.user_input import DiscreteGrid, LinspaceGrid, LogspaceGrid, Model
 
 __all__ = ["mark", "Model", "LinspaceGrid", "LogspaceGrid", "DiscreteGrid"]
