+++ tests/test_simulate.py
@@ -160,7 +160,6 @@ def test_simulate_using_get_lcm_function(
     assert_array_equal(res.loc[last_period_index, :]["retirement"], 1)
 
     for period in range(n_periods):
-
         # assert that higher wealth leads to higher consumption in each period
         assert (res.loc[period]["consumption"].diff()[1:] >= 0).all()
 
@@ -255,7 +254,6 @@ def test_effect_of_disutility_of_work():
     # Asserting
     # ==================================================================================
     for period in range(5):
-
         # We expect that individuals with lower disutility of work, work (weakly) more
         # and thus consume (weakly) more
         assert (
